Modify structure of CGenNodeTimesync (and other)
CGenNodeTimesync was not consistent with CGenNodeStateless: m_port_id was
packed in a different place of the structure which resulted with core dumps
in PerPortProfile::stop_traffic.

Changes to the CTimesyncEngine's master part.  They now generate a flow of
packages required to do the transmission.

Changes from Artur's branch that set up the MAC addresses.

TODO: m_mac_addr MUST BE moved higher in the CGenNodeTimesync struct (same place
as CGenNodePCAP for example).